update white list conditionally in config helper
